refactor: rename to oscd-tree-grid
