+++ README.head.md
@@ -1,21 +1,21 @@
-# \<oscd-tree-table>
+# \<oscd-tree-grid>
 
 This webcomponent follows the [open-wc](https://github.com/open-wc/open-wc) recommendation.
 
 ## Installation
 
 ```bash
-npm i oscd-tree-table
+npm i oscd-tree-grid
 ```
 
 ## Usage
 
 ```html
 <script type="module">
-  import 'oscd-tree-table';
+  import 'oscd-tree-grid';
 </script>
 
-<oscd-tree-table></oscd-tree-table>
+<oscd-tree-grid></oscd-tree-grid>
 ```
 
 ## Linting and formatting
