+++ oscd-tree-grid.test.ts
@@ -3,8 +3,8 @@ import { html } from 'lit';
 
 import { visualDiff } from '@web/test-runner-visual-regression';
 
-import './oscd-tree-table.js';
-import type { OscdTreeTable } from './oscd-tree-table.js';
+import './oscd-tree-grid.js';
+import type { OscdTreeTable } from './oscd-tree-grid.js';
 
 const factor = process.env.CI ? 2 : 1;
 
@@ -16,4 +16,4 @@ function timeout(ms: number) {
 
 mocha.timeout(2000 * factor);
 
-describe('oscd-tree-table', () => {});
+describe('oscd-tree-grid', () => {});
